+++ src/main/factories/signup/signup-validation.ts
@@ -1,6 +1,6 @@
 import { ValidationComposite } from '../../../presentation/helpers/validators/validation-composite'
 import { RequiredFieldValidation } from '../../../presentation/helpers/validators/required-field-validation'
-import { Validation } from '../../../presentation/helpers/validators/validation'
+import { Validation } from '../../../presentation/protocols/validation'
 import { CompareFieldsValidation } from '../../../presentation/helpers/validators/compare-fields-validation'
 import { EmailValidatorAdapter } from '../../../utils/email-validator-adapter'
 import { EmailValidation } from '../../../presentation/helpers/validators/email-validation'
