+++ src/presentation/controllers/login/login-protocols.ts
@@ -1,3 +1,3 @@
 export * from '../../protocols'
-export * from '../../helpers/validators/validation'
+export * from '../../protocols/validation'
 export * from '../../../domain/use-cases/authentication'
