+++ test/login-validation.spec.ts
@@ -1,7 +1,7 @@
 import { makeLoginValidation } from '../src/main/factories/login/login-validation'
 import { EmailValidation } from '../src/presentation/helpers/validators/email-validation'
 import { RequiredFieldValidation } from '../src/presentation/helpers/validators/required-field-validation'
-import { Validation } from '../src/presentation/helpers/validators/validation'
+import { Validation } from '../src/presentation/protocols/validation'
 import { ValidationComposite } from '../src/presentation/helpers/validators/validation-composite'
 import { EmailValidator } from '../src/presentation/protocols/email-validator'
