+++ test/signup-validation.spec.ts
@@ -2,7 +2,7 @@ import { makeSignUpValidation } from '../src/main/factories/signup/signup-valida
 import { CompareFieldsValidation } from '../src/presentation/helpers/validators/compare-fields-validation'
 import { EmailValidation } from '../src/presentation/helpers/validators/email-validation'
 import { RequiredFieldValidation } from '../src/presentation/helpers/validators/required-field-validation'
-import { Validation } from '../src/presentation/helpers/validators/validation'
+import { Validation } from '../src/presentation/protocols/validation'
 import { ValidationComposite } from '../src/presentation/helpers/validators/validation-composite'
 import { EmailValidator } from '../src/presentation/protocols/email-validator'
